removed models and repo folder; moved User, UserDTO and UserRepository into main package, changed LocalDateTime to String because dynamodb can't handle time
