+++ src/main/java/com/revature/post_user/PostUserHandler.java
@@ -7,9 +7,6 @@ import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyRequestEvent;
 import com.amazonaws.services.lambda.runtime.events.APIGatewayProxyResponseEvent;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
-import com.revature.models.User;
-import com.revature.models.UserDTO;
-import com.revature.repos.UserRepository;
 
 public class PostUserHandler implements RequestHandler<APIGatewayProxyRequestEvent, APIGatewayProxyResponseEvent> {
